fixed code climate issue in utilities
